moved partnerService class and PartnerInventoryService into partner package
