+++ BookStore/src/com/ebook/model/partner/PartnerInventory.java
@@ -1,20 +1,11 @@
 package com.ebook.model.partner;
 
-import java.util.HashMap;
+
 import java.util.Map;
 
 import com.ebook.model.item.Product;
 
 public class PartnerInventory{
 
-    // Integer value is for keeping track of remain quantity of the product.
-    Map <Product,Integer> partnerIntentory = new HashMap<>();
-
-    public void addProduct(Product product, int quantity) {
-
-    }
-
-
-
-
+ 
 }
